+++ src/main/kotlin/csense/idea/base/bll/kotlin/KtBlockExpression.kt
@@ -10,6 +10,7 @@ import org.jetbrains.kotlin.psi.KtBlockExpression
  * @return (com.intellij.psi.PsiElement..com.intellij.psi.PsiElement?)
  */
 fun KtBlockExpression.addFirstInScope(element: PsiElement) = addAfter(element, lBrace)
+
 /**
  * Adds the given element as the last element inside of this block (so not outside the scope)
  * @receiver KtBlockExpression